Store a memo that the user will add to the transaction, instead of creating a new bnb wallet.
This is because transferring funds from one bnb wallet to another requires BNB for tx fees.
Thus trying to consolidate all the bnb from the created wallets is very costly.

We replace this with a memo instead, which is a randomly generated string.
We use this to identify who the transaction belongs to.